Refactor : client/config의 view에 표시될 메시지 상수들을 분리
상세 :
- client/constants/message에서 메시지 관련 상수들 관리

 #388
