+++ client/src/service/ClientManager.js
@@ -7,7 +7,8 @@ import ChattingManager from './ChattingManager';
 import { browserLocalStorage, makeViewPlayerList } from '../utils';
 import EVENTS from '../constants/events';
 import actions from '../actions';
-import { GAME_END_SCOREBOARD_TITLE, SOCKETIO_SERVER_URL } from '../config';
+import { GAME_END_SCOREBOARD_TITLE } from '../constants/message';
+import { SOCKETIO_SERVER_URL } from '../config';
 
 class ClientManager {
   constructor({ history, roomIdFromUrl, isPrivateRoomCreation }) {
